changes to homepage
I worked adding color to the home page. added a box around the why choose us description. the box has a ton of padding on the left, making it unique. the padding changes to adjust to 3 media queries that i've added at the top of index2.html in the style section. They are listed from biggest screen size to smallest.